requirements.txt
PyQt6
discord.py
sqlalchemy